+++ src/Phobos.h
@@ -84,7 +84,7 @@ public:
 		static bool ToolTipBlur;
 		static bool PrioritySelectionFiltering;
 		static bool PriorityDeployFiltering;
-		static bool ModernControls;
+		static bool RightClickCommand;
 		static bool TypeSelectUseIFVMode;
 		static bool DevelopmentCommands;
 		static bool SuperWeaponSidebarCommands;